// src/models/filter-options.interface.ts
export enum FilterOptions {
    EVERYTHING = "everything",
    TOP_HEADLINES = "top-headlines",
    COUNTRY = "country",
    CATEGORY = "category",
    SOURCE = "source",
    LANGUAGE = "language",

}

export enum EndpointOption {
    EVERYTHING = "everything?",
    TOP_HEADLINES = `top-headlines?`,

}